requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Financial Data
yfinance>=0.2.30
pandas-datareader>=0.10.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Data Storage
pyarrow>=13.0.0
fastparquet>=2023.8.0

# Database - PostgreSQL/TimescaleDB
psycopg2-binary>=2.9.7
SQLAlchemy>=2.0.0

# Statistical Modeling
statsmodels>=0.14.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# MLflow & Experiment Tracking
mlflow>=2.18.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
networkx>=3.1
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# FUTURE: News & Sentiment Analysis (commented out for now)
# feedparser>=6.0.10
# newspaper3k>=0.2.8
# beautifulsoup4>=4.12.2
# lxml>=4.9.3
# nltk>=3.8.0
# transformers>=4.30.0
# pymongo>=4.5.0
# python-dateutil>=2.8.2